requirements.txt
# Digital Employee - Silver Tier Dependencies
# Install with: pip install -r requirements.txt

# =============================================================================
# CORE DEPENDENCIES (Required)
# =============================================================================

# Environment variable management
python-dotenv>=1.0.0

# File system watching
watchdog>=3.0.0

# HTTP requests (for LinkedIn API)
requests>=2.28.0

# =============================================================================
# GMAIL WATCHER (Required for email monitoring)
# =============================================================================

# Google Gmail API
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-api-python-client>=2.0.0

# =============================================================================
# WHATSAPP WATCHER (Optional - Twilio integration)
# =============================================================================

# Twilio API for WhatsApp
twilio>=8.0.0

# =============================================================================
# LINKEDIN MCP (Optional - LinkedIn publishing)
# =============================================================================

# Already included: requests>=2.28.0

# =============================================================================
# LLM ROUTER (Required for AI reasoning)
# =============================================================================

# LiteLLM - Unified interface for 100+ LLM providers
litellm>=1.60.0

# =============================================================================
# SCHEDULING & AUTOMATION (Optional)
# =============================================================================

# Scheduler for automated runs
schedule>=1.2.0

# =============================================================================
# OPTIONAL ENHANCEMENTS
# =============================================================================

# YAML parsing for enhanced config
# pyyaml>=6.0

# Rich terminal output
# rich>=13.0.0

# =============================================================================
# INSTALLATION COMMANDS
# =============================================================================

# Full installation (all features):
#   pip install -r requirements.txt

# Minimal installation (core only):
#   pip install python-dotenv watchdog requests litellm

# Production installation (no optional):
#   pip install python-dotenv watchdog requests google-auth google-auth-oauthlib google-api-python-client twilio schedule litellm


playwright
